:cake: Initial commit
This initial commit includes a web-pack config that builds a bundled
javascript file for the inject code.

The bundled `inject.bundled.js` is then added to every page as a
content script in the extensions manifest file.

The webpack config also passes babel-polyfill into the entry array
meaning es6 can be used.