Add storage sync spike
This commit adds a basic implementation of using local storage via
chrome.storage.

It also moves the spike component out of of inject.jsx to pave the
way for how future components/containers will be created.

Ideally, containers should just be used to inject a set of components
into the page somewhere.