Add placeholder basics for EventFollowing class
This class will allow users to follow events in future. Followed
events will be listed in a "Following" drop-down on all pages
and will allow quick access back to the events.

Future features to the following drop-down could also be an update
ticker that goes through each followed event and updates the stats
for that event, and also linking the event with your bet-slip to
show how many bets you've won/lost for that event, including
profits and losses.

* Add bare-bones serializer for Football event.