Port followEvent to use redux for state manangement
In this commit I've removed a lot of code that previously used
chrome.storage and have replaced it with redux to keep track of the
state.

The followEvent reducer is passed into components as the store prop,
the component then sends messages to the store to update the global
state of the followed list.

The component also subscribes to the store to re-render itself
whenever changes to the data are made. This means that we reference
data from the store rather than using react state.

The reducer also contains a sync action that updates localstorage
for the key.

It also attempts to load default state from localstorage, otherwise
defaulting to a blank object.